+++ python/metatomic_torchsim/metatomic_torchsim/__init__.py
@@ -1,0 +1,4 @@
+from ._model import MetatomicModel
+
+
+__all__ = ["MetatomicModel"]